requirements.txt
Flask_SQLAlchemy==2.4.0
scikit_image==0.15.0
scikit_learn==0.21.3
opencv_python_headless==4.1.1.26
scipy==1.3.0
requests==2.22.0
SQLAlchemy==1.3.5
torch==1.5.0
torchvision==0.6.0
Flask_Restless==0.17.0
numpy==1.16.4
Flask==1.0.3
umap_learn==0.3.10
Pillow==6.2.0
tensorboardX==1.9
ttach==0.0.2
albumentations==0.4.3
config==0.4.2
